Move manifest processing functions to mendel-development
